pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nps-relacional-sellers"
version = "0.1.0"
description = "Modelo de análisis NPS Relacional para Sellers de Mercado Pago"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-bigquery-storage>=2.25.0",
    "google-auth>=2.29.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.4",
    "matplotlib>=3.8.0",
    "pydantic>=2.7.0",
    "python-dateutil>=2.9.0",
    "tqdm>=4.66.0",
    "click>=8.1.7",
    "db-dtypes>=1.5.0",
    "pyarrow>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
]
pdf = [
    "weasyprint>=62.0",  # Para generar PDF desde HTML
]
llm = [
    "google-generativeai>=0.7.0",  # Gemini SDK
]

[project.scripts]
nps-model = "nps_model.__main__:cli"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # line too long (handled by black)

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
